Return wall health, paddle movement and brick shooter script update
Added script to create a health system for the return wall to take damage and be able to lose the game.
Changed player/paddle movement from mouse input to a and d key inputs. Also changed dotted aim line movement from a and d keys to q and e keys.
Created a script and prefab for brick shooter that can shoot balls.