Trajectory line, coin pickups and ball generator
Created a trajectory dotted line that gets smaller towards the end of it to replace the short aim line. Made it dynamic so I can change the length and number of bounces it will show.
Created coins that enemies/bricks drop and the player can collect them by clicking with the left mouse button over them.
Made a ball generator with a fill bar so once it fills up over a period of time, a new ball is created so you don't get stuck on the play screen if you lose all of your balls.
Created a button to upgrade the ball generator that decreases the time it takes to make a new ball and increases in cost each upgrade.