Ball shooting formations, fire mode buttons and ball generator tweak
Created different formations/shooting types, single, burst, horizontal, circle and triangle, with buttons that are toggled to change them.
Circle and triangle are not working exactly how I want them to.
Made it so the ball generator progress stops and goes back to 0 when the level is complete. Then after having transitioned to a new level, progress for the next ball to be generated is still at 0 and starts up again.