Player and enemy ball health, speed and material update
Created separate materials for player and enemy balls that get darker the lower their health is.
Made it so the player balls health is set tp full and their speed is increased by 5 when they bounce off the player/paddle.
Made it so player balls have a scale pop effect when they hit the player.